feat(electron): delete old AudioSession/EqBridge/FfmpegToolchain, consolidate into DaemonClient
Delete 3 legacy audio module files no longer needed since the ECHO Audio Daemon handles all audio processing:
- AudioSession.ts (thin wrapper around DaemonClient)
- EqBridge.ts (EQ state manager, moved into DaemonClient)
- FfmpegToolchain.ts (FFmpeg binary resolution, moved into DaemonClient)

All exports from the deleted files are consolidated into DaemonClient.ts.
All ~50 import sites across src/ are updated to point to DaemonClient.ts.
npx tsc --noEmit passes with zero errors.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
