+++ src/main/integrations/lastfm/LastFmService.test.ts
@@ -10,7 +10,7 @@ vi.mock('../../app/appSettings', () => ({
   setAppSettings: vi.fn(),
 }));
 
-vi.mock('../../audio/AudioSession', () => ({
+vi.mock('../../audio/DaemonClient', () => ({
   getAudioSession: () => ({
     getStatus: vi.fn(),
   }),
