+++ src/main/library/LibraryServiceGroupingRefresh.test.ts
@@ -19,7 +19,7 @@ vi.mock('electron', () => ({
 
 let playbackState: 'idle' | 'loading' | 'playing' | 'paused' | 'stopped' = 'idle';
 
-vi.mock('../audio/AudioSession', () => ({
+vi.mock('../audio/DaemonClient', () => ({
   getAudioSession: () => ({
     getStatus: () => ({
       state: playbackState,
